Agregada la restricción de no poder enviar la orden de compra antes de que se termine de actualizar el valor
